backend/requirements.txt
cat > requirements.txt << 'EOF'
fastapi==0.111.0
uvicorn[standard]==0.29.0
anthropic==0.28.0
httpx==0.27.0
sqlalchemy==2.0.30
python-dotenv==1.0.1
pydantic==2.7.1
python-multipart==0.0.9
EOF